Pull out LockAndTimestampServiceFactory
